move logError to Utils
