+++ src/error.js
@@ -25,8 +25,6 @@
 * for the JavaScript code in this file.
 *
 */
-import {createLogger} from './utils'
-const logger = createLogger();
 
 export default class extends Error {
   constructor(status, payload, ...params) {
@@ -34,18 +32,4 @@ export default class extends Error {
     this.status = status;
     this.payload = payload;
   }
-}
-
-export function logError(err) {
-  if (err instanceof ApiError) {
-    logger.log('error', JSON.stringify(err));
-    return;
-  }
-
-  if (err === 'SIGINT') {
-    logger.log('error', err);
-    return;
-  }
-
-  logger.log('error', err.stack === undefined ? err : err.stack);
 }
